zeromq/requirements.txt
numpy
pyzmq
mpi4py
pyyaml
